+++ src/RBF/Rbf.hpp
@@ -6,23 +6,23 @@
 #include "Debugging/Log.hpp"
 #include <vector>
 
-double distance(glm::vec3 a, glm::vec3 b);
-double multiQuadra(double x);
-double invMultiQuadra(double x);
-double gaussian(double x);
+float distance(glm::vec3 a, glm::vec3 b);
+float multiQuadra(float x);
+float invMultiQuadra(float x);
+float gaussian(float x);
 
 
 
 class RBF{
 public:
-	RBF(std::vector<glm::vec3> anchorPts, Eigen::VectorXd valuesAtAnchorPts, std::function<double(double)> phi);
+	RBF(std::vector<glm::vec3> anchorPts, Eigen::VectorXf valuesAtAnchorPts, std::function<double(double)> phi);
 
-	double eval(const glm::vec3& pos);
+	float eval(const glm::vec3& pos);
 
 
 private:
-	void calculOmega(Eigen::VectorXd valuesAtAnchorPts);
+	void calculOmega(Eigen::VectorXf valuesAtAnchorPts);
 	std::vector <glm::vec3> m_anchorPts;
-	Eigen::VectorXd m_omega;
+	Eigen::VectorXf m_omega;
 	std::function<double(double)> m_phi;
 };
